Enforce float32 for Image gray/detect_mat luminance layers
gray and detect_mat were declared float32 but silently ran float64:
skimage.rgb2gray returns float64 and _set_from_matrix assigned it
verbatim, so the declared contract was never enforced. Enforce it via
ImageData.__setattr__, halving the RAM and on-disk footprint of the two
largest image layers (~21% smaller HDF5 files) at negligible precision
cost (float32 ~7 significant digits, ~500x finer than the uint16
quantization step). The docs and test_rgb_modes_produce_float32 already
documented/expected float32; this aligns runtime with that contract.

To keep measurements numerically faithful on the now-float32 intensity
layers, reductions accumulate in float64: a shared _as_accumulation_dtype
guard upcasts inputs in the MeasureFeatures sum/mean/stddev/variance/
median/q1/q3/iqr helpers, and the regionprops intensity_image sites in
MeasureBounds / MeasureSymmetricZones cast up too. Measurement columns
stay float64 (schema-preserving).

Validated on the FilamentousFungiPipeline (real tutorial image): a
float64-vs-float64 control is bit-identical, while float32 drifts
segmentation by 10/1,461,952 px (object count unchanged, per-object IoU
>=0.9999) and measurements by <=0.2%. Updated three tests that pinned
bit-exact float64 to float32/allclose and added a float32 contract +
HDF5 round-trip test class.

Out of scope: the manual tests/migration/ pydantic-equivalence goldens
encode bit-exact float64 detect_mat; they are not in testpaths and never
run in CI, so they do not gate this change (would need regeneration if
ever re-run).

Verification: full unit+smoke+integration suite green (3595 passed, 13
skipped); measure + abc_ + HDF5 round-trip green with slow enabled (69
passed); mypy shows only pre-existing debt; ruff clean on changed files.

+++ src/phenotypic/enhance/_enhance_features.py
@@ -195,8 +195,9 @@ class EnhanceFeatures(ImageEnhancer):
         output_map = {"M": result.M, "m": result.m, "pc_sum": result.pc_sum}
         selected = output_map[self.output]
 
-        # Ensure output is in [0, 1] range for detect_mat compatibility
-        image.detect_mat[:] = np.clip(selected, 0.0, 1.0).astype(np.float64)
+        # Ensure output is in [0, 1] range for detect_mat compatibility.
+        # detect_mat enforces float32 on assignment, so no explicit cast is needed.
+        image.detect_mat[:] = np.clip(selected, 0.0, 1.0)
         return image
 
     def _phasecong3(self, img: np.ndarray) -> _PhaseCong3Result:
